+++ tests/unit/small_text/integrations/transformers/classifiers/test_classification.py
@@ -350,4 +350,4 @@ class TestTransformerBasedClassification(unittest.TestCase):
 
         proba = clf.predict_proba(test_set)
         self.assertEqual(0, proba.shape[0])
-        self.assertTrue(np.issubdtype(proba.dtype, np.float))
+        self.assertTrue(np.issubdtype(proba.dtype, float))
